+++ src/hooks/index.ts
@@ -27,7 +27,7 @@ export { useOffsetPosition } from './useOffsetPosition';
 export { useOrientation } from './useOrientation';
 export { useOverlap } from './useOverlap';
 export { useParticles } from './useParticles';
-export { usePhysicsBody, useBoxPhysics, useCirclePhysics } from './usePhysics';
+export { usePhysicsBody, useBoxPhysics, useCirclePhysics, useSyncPositions } from './usePhysics';
 export { usePhysicsEngine, usePhysicsCollision } from './usePhysicsEngine';
 export { usePlatformMovement } from './usePlatformMovement';
 export { usePointer } from './usePointer';